+++ lib/pgvector.dart
@@ -1,5 +1,6 @@
 export 'src/bit.dart' show Bit;
 export 'src/halfvec.dart' show HalfVector;
 export 'src/pgvector.dart' show pgvector;
+export 'src/postgres.dart' show pgvectorEncoder;
 export 'src/sparsevec.dart' show SparseVector;
 export 'src/vector.dart' show Vector;