+++ test/postgres_test.dart
@@ -11,7 +11,9 @@ void main() {
             port: 5432,
             database: 'pgvector_dart_test',
             username: Platform.environment['USER']),
-        settings: ConnectionSettings(sslMode: SslMode.disable));
+        settings: ConnectionSettings(
+            sslMode: SslMode.disable,
+            typeRegistry: TypeRegistry(encoders: [pgvectorEncoder])));
 
     await connection.execute('CREATE EXTENSION IF NOT EXISTS vector');
     await connection.execute('DROP TABLE IF EXISTS items');
@@ -23,25 +25,25 @@ void main() {
         Sql.named(
             'INSERT INTO items (embedding, half_embedding, binary_embedding, sparse_embedding) VALUES (@embedding1, @half_embedding1, @binary_embedding1, @sparse_embedding1), (@embedding2, @half_embedding2, @binary_embedding2, @sparse_embedding2), (@embedding3, @half_embedding3, @binary_embedding3, @sparse_embedding3)'),
         parameters: {
-          'embedding1': Vector([1, 1, 1]).toString(),
-          'embedding2': Vector([2, 2, 2]).toString(),
-          'embedding3': Vector([1, 1, 2]).toString(),
-          'half_embedding1': HalfVector([1, 1, 1]).toString(),
-          'half_embedding2': HalfVector([2, 2, 2]).toString(),
-          'half_embedding3': HalfVector([1, 1, 2]).toString(),
-          'binary_embedding1': '000',
-          'binary_embedding2': '101',
-          'binary_embedding3': '111',
-          'sparse_embedding1': SparseVector([1, 1, 1]).toString(),
-          'sparse_embedding2': SparseVector([2, 2, 2]).toString(),
-          'sparse_embedding3': SparseVector([1, 1, 2]).toString()
+          'embedding1': Vector([1, 1, 1]),
+          'embedding2': Vector([2, 2, 2]),
+          'embedding3': Vector([1, 1, 2]),
+          'half_embedding1': HalfVector([1, 1, 1]),
+          'half_embedding2': HalfVector([2, 2, 2]),
+          'half_embedding3': HalfVector([1, 1, 2]),
+          'binary_embedding1': Bit([false, false, false]),
+          'binary_embedding2': Bit([true, false, true]),
+          'binary_embedding3': Bit([true, true, true]),
+          'sparse_embedding1': SparseVector([1, 1, 1]),
+          'sparse_embedding2': SparseVector([2, 2, 2]),
+          'sparse_embedding3': SparseVector([1, 1, 2])
         });
 
     List<List<dynamic>> results = await connection.execute(
         Sql.named(
             'SELECT id, embedding, binary_embedding, sparse_embedding FROM items ORDER BY embedding <-> @embedding LIMIT 5'),
         parameters: {
-          'embedding': Vector([1, 1, 1]).toString()
+          'embedding': Vector([1, 1, 1])
         });
     expect(results.map((r) => r[0]), equals([1, 3, 2]));
     expect(Vector.fromBinary(results[1][1].bytes), equals(Vector([1, 1, 2])));