+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.1 (unreleased)
+## 0.2.0 (unreleased)
 
 - Added `Vector`, `HalfVector`, `Bit`, and `SparseVector` classes